Move deprecated APIs to sub-folder
And remove inline documentation.  These APIs are deprecated no
need to document other than replacement.
