+++ src/runtime_src/core/include/deprecated/CMakeLists.txt
@@ -1,0 +1,9 @@
+set(XRT_DEPRECATED_HEADER_SRC
+  xrt.h)
+
+install (FILES ${XRT_DEPRECATED_HEADER_SRC} DESTINATION ${XRT_INSTALL_INCLUDE_DIR}/deprecated)
+
+message("-- XRT deprecated header files")
+foreach (header ${XRT_DEPRECATED_HEADER_SRC})
+  message("-- ${header}")
+endforeach()
